+++ src/main/java/org/c4marathon/assignment/model/type/MailStatus.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.domain.model;
+package org.c4marathon.assignment.model.type;
 
 public enum MailStatus {
     PENDING,
