+++ src/main/java/org/c4marathon/assignment/repository/MailLogJpaRepository.java
@@ -1,7 +1,7 @@
 package org.c4marathon.assignment.repository;
 
 import org.c4marathon.assignment.domain.MailLog;
-import org.c4marathon.assignment.domain.model.MailStatus;
+import org.c4marathon.assignment.model.type.MailStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
